Refactors the scan_to_pointcloud node into a header and cpp file
