+++ igvc_gazebo/nodes/scan_to_pointcloud/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(scan_to_pointcloud main.cpp)
+add_executable(scan_to_pointcloud scan_to_pointcloud.cpp scan_to_pointcloud.h)
 add_dependencies(scan_to_pointcloud ${catkin_EXPORTED_TARGETS})
 target_link_libraries(scan_to_pointcloud ${catkin_LIBRARIES} ${PCL_LIBRARIES})
